Don't use basename from other Tempfiles
Currently if we do

  size_800 = resize_to_limit(image,    800, 800)
  size_500 = resize_to_limit(size_500, 500, 500)
  size_300 = resize_to_limit(size_300, 300, 300)

The basename of each previous filename will be added to the next
path of the next file, making the file path grow with useless
information.

We actually don't want any part of the given file to be used when
copying it into a Tempfile (just the file extension), so we remove this
behaviour.