Vips: Apply only saver options that are defined
Different Vips savers accept different options. At the moment the
developer needs to know exactly which formats accept provided saver
options, because when these options are forwarded to a saver which
doesn't accept this options an error is raised.

Instead of the developer having to know which output formats accept the
:Q option and which don't, they can now specify this option and it will
be forwarded only to savers that accept it, while for others it will be
ignored.

See #17 for the discussion.