[minimagick] Add loader and saver options
ImageMagick does have options that apply specifically to loading and
saving images. The saver options could already be defined via the
chainable API, but the loader options couldn't, because they need to be
specified before the input image in the shell command.

While we're here, we also recommend specifying the "jpeg:size" hint when
generating multiple thumbnails to avoid loading the whole image into
memory, which resulted in about 2x speedup in my benchmarks.