Switch #resize_* operations back to "-resize"
Using "-thumbnail" is faster in certain cases, specifically with large
images and TIFFs, and it's recommended in the ImageMagick documentation.

However, unlike "-resize", "-thumbnail" strips away data, namely EXIF,
GPS, and TIFF tags, including any attached profiles except color
profiles. This can lead to surprising behaviour, and people might
be relying on this information to be kept, so to follow the principle of
least surprise we're reverting back to "-resize".

We'll probably add a way for people to opt-in to use "-thumbnail",
though it might be unnecessary once we add automatic JPEG hints for
loading only the subset of image pixels needed for processing.