[vips] Don't apply #autorot if :autorotate option was passed in
When the user passed the :autorotate loader option, we want to move out
of the way and not call Vips::Image#autorot. If :autorotate was true,
then there is no need for the additional #autorot. If :autorotate was
false, then we definitely don't want to autorotate even if the user
specified not to.

In general this will allow users to be able to use :autorot and
:autorotate interchangeably, which will definitely prevent potential
confusion if they accidentally used :autorotate instead of :autorot.

I still want to keep the :autorot option, because I want to make it
explicit that this option will make an additional Vips::Image#autorot.
That call is useful because then the image will be autorotated even when
a Vips::Image was given as the source.