[core] Rename Builder#default_options to Builder#options
Having builder options reader named #default_options didn't make sense
in the first place, it was only forced by the design of the chainable
API. So we refactor the chainable API to allow the reader method to be
called #options. We also freeze the options to ensure they won't be
mutated.

This also removes the @default_options instance variable added to the
ImageProcessing::MiniMagick or ImageProcessing::Vips module, as well as
the #default_options reader method.

While we're here, we also document retrieving builder options, as
someone already asked about this before.