Depend on mini_magick and ruby-vips gems
This will make maintenance a lot easier, because we'll be able to adapt to
changes in mini_magick and ruby-vips gems, as we will have the ability
to bump their versions.

This also makes installation easier, because the user doesn't need to
remember to add those gems to the Gemfile, and doesn't need to remember
which versions should be specified.

The downside is that both gems will always be installed, which might
cause concern for some users. However, considering that both gems are
very small (mini_magick is 43KB and ruby-vips is 233KB), and that both
can be normally installed without ImageMagick or libvips being
installed, I think the ends justify the means. And there is no rational
reason for concern, because only the gem that's used will actually be
loaded.