[core] Extend #apply to accept any builder commands
Previously only operations were allowed in #apply. This commit extends
it to accept any builder commands, which includes :convert, :loader,
:saver etc. This will be useful for Active Storage, as it will allow
users to add conversion, loader and saver options to their on-the-fly
transformations.

This change also allows values given to #apply to be Hashes, which
previously wasn't possible in case someone wanted to use it for a
certain libvips operation.

This also brings back MiniMagick's support for "+" operators, which we
now document as well.