Initialize processor with the accumulator object
This way we don't have to pass the accumulator object (MiniMagick::Tool,
Vips::Image) to each processor method. This will make it easier to read
the source code, because the arguments now match the caller uses. We
also rename it "accumulator", as something that describes both
Vips::Image and MiniMagick::Tool objects.

We never even used the pipeline information so far, and I think it's bad
that we'd be coupling Processor and Pipeline classes this way. So we
remove it.

We move Processor#load_image and Processor#save_image to class methods,
as they differ from other processing methods (e.g. #save_image doesn't
return a Vips::Image object). We also move Processor#apply_operation to
a class method. That way in Pipeline we're interacting only with the
Processor class.

We also require each Vips processing methods (except #set, #set_value,
and #set_type) to return the Vips::Image object. That should make more
difficult to accidentally forget to return the Vips::Image object that
we've applied an operation to.