[minimagick] Add :loader for explicitly setting input file type
By default, ImageMagick guesses the input file type based on magic
headers and other information. However, it's not always possible for
ImageMagick to correctly identify the file based on its content, and in
some cases it can be a security flaw. So, ImageMagick gives you the
ability to explicitly set the input file type.

  convert jpg:input.jpg ...

The vips backend already provides the equivalent :loader option, so we
use the same name for the minimagick counterpart as well.

  ImageProcessing::MiniMagick
    .source("input.jpg")
    .loader(loader: "jpg")
    # ...